requirement.txt
bing-image-downloader
numpy
Pillow
opencv-python
matplotlib
scipy
split-folders